default dtype to float32

+++ xinfer/transformers/auto.py
@@ -14,7 +14,7 @@ from ..base_model import BaseModel
 
 class Vision2SeqModel(BaseModel):
     def __init__(
-        self, model_id: str, device: str = "auto", dtype: str = "float16", **kwargs
+        self, model_id: str, device: str = "auto", dtype: str = "float32", **kwargs
     ):
         device = "cuda" if device == "auto" and torch.cuda.is_available() else "cpu"
